Initial re-factor of mavlink.
